+++ serde_arrow/src/test_end_to_end/issue_90.rs
@@ -9,10 +9,7 @@ use crate::{
     schema::{SchemaLike, TracingOptions},
 };
 
-use crate::_impl::arrow::{
-    _raw::{array::RecordBatch, schema::Schema},
-    datatypes::Field,
-};
+use crate::_impl::arrow::{_raw::schema::Schema, array::RecordBatch, datatypes::Field};
 
 #[derive(Deserialize, Serialize, Debug, PartialEq, Clone)]
 pub struct Distribution {